undo refactore of pytorch_no_op fixture and add symlinks instead
