+++ e2e_tests/tests/fixtures/pytorch_no_op/const.yaml
@@ -1,0 +1,19 @@
+name: noop-pytorch
+data:
+  model_type: single_output
+entrypoint: model_def:NoopPyTorchTrial
+hyperparameters:
+  global_batch_size: 32
+  dataset_len: 64
+scheduling_unit: 1
+searcher:
+  name: single
+  metric: validation_error
+  max_length:
+    batches: 3
+  smaller_is_better: true
+max_restarts: 0
+min_checkpoint_period:
+  batches: 1
+min_validation_period:
+  batches: 1
